Support Multiple Contract Configurations (#6)
This PR adds support for additional chains where the Safe address is
computed in the same way as Ethereum. For this, we add a `--chain`
configuration for specifying a new chain. Since v1.4.1 is only released
on Mainnet and Gnosis Chain, only those two chains are supported for
now.

Additionally, there are new parameters to completely override all
chain-specific data (contract addresses and init codes) as an
"officially unsupported escape hatch".