WASM Bundle With Demo App (#8)
Fixes #5

This PR adds a WASM bundle with an (_unpublished_) NPM `package.json` so
that it can be used on the Web.

An example web page is included in the `wasm/example` sub-directory (see
`wasm/README.md` for instructions).

As a design choice, the source is bundled into a single file, which
includes base-64 encoded WASM binary blob and uses `Blob` URLs for the
worker. This allows the worker to be bundled without requiring
additional HTTP requests for additional resources.