Safe v1.4.1 With Multi-Chain Support (#13)
This is a big one...

This PR updates the contracts to use the latest v1.4.1 version, as well
as adding support for the `SafeToL2Setup` multi-chain support mechanism.
This allows Safes generated with `deadbeef` to be replayed on different
networks and keep the same address.

In addition, the default chains were expanded to include all supported
chains from the Safe{Wallet} interface (with the exception of zkSync
Era).